requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1
beanie==1.15.4
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
debugpy==1.6.4
Django==4.1.4
dnspython==2.2.1
exceptiongroup==1.0.2
filelock==3.8.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
kaitaistruct==0.10
lxml==4.9.1
motor==3.1.1
multidict==6.0.2
nest-asyncio==1.5.6
outcome==1.2.0
packaging==21.3
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.6
pymongo==4.3.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
selenium==4.6.0
selenium-requests==2.0.2
selenium-wire==5.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tldextract==3.4.0
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
wsproto==1.2.0
-e git+ssh://github-personal/nvvu99/xss-scanner.git@14fcd5c710e6d309b03df872213c339bbb9011a3#egg=xss_scanner
yarl==1.8.1
zope.interface==5.5.2
zstandard==0.19.0
